Did some file reorganization around imagesets.
