+++ runner/train.py
@@ -10,9 +10,9 @@ from omegaconf import DictConfig
 from torch.optim.optimizer import Optimizer
 from torch.utils.data import DataLoader
 
+from retinal_rl.classification.imageset import Imageset
 from retinal_rl.classification.loss import ClassificationContext
 from retinal_rl.classification.training import process_dataset, run_epoch
-from retinal_rl.dataset import Imageset
 from retinal_rl.models.brain import Brain
 from retinal_rl.models.objective import Objective
 from runner.analyze import analyze
